backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
pytest==7.2.2
pytest-django==4.5.2
black==22.3.0
flake8==4.0.1
mypy==0.910
django-environ==0.8.1